+++ packages/executor/src/actions/clearTxIntentions.ts
@@ -1,0 +1,6 @@
+import type { MidlContextState } from "@midl-xyz/midl-js-react";
+import type { StoreApi } from "zustand";
+
+export const clearTxIntentions = (store: StoreApi<MidlContextState>) => {
+	store.setState({ intentions: [] });
+};