+++ packages/executor/src/utils/calculateTransactionCost.test.ts
@@ -1,3 +1,5 @@
+// @vitest-environment node
+
 import { describe, expect, it } from "vitest";
 import { midlConfig } from "~/__tests__/midlConfig";
 import { calculateTransactionsCost } from "~/utils/calculateTransactionsCost";