+++ packages/executor/src/utils/index.ts
@@ -7,3 +7,4 @@ export * from "./getEVMFromBitcoinNetwork";
 export * from "./calculateTransactionsCost";
 export * from "./convertBTCtoETH";
 export * from "./convertETHtoBTC";
+export * from "./viem2ethers";