+++ apps/e2e/CHANGELOG.md
@@ -1,5 +1,14 @@
 # e2e
 
+## 1.0.45
+
+### Patch Changes
+
+- Updated dependencies
+  - @midl-xyz/midl-js-executor@0.0.73
+  - @midl-xyz/midl-js-core@0.0.37
+  - @midl-xyz/midl-js-react@0.0.44
+
 ## 1.0.44
 
 ### Patch Changes